src/Hybrid ResNet Lite/requirements.txt
absl-py==0.9.0
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Cython==0.25.2
decorator==4.4.1
defusedxml==0.6.0
docutils==0.16
entrypoints==0.3
filelock==3.0.12
google-auth==1.11.0
google-auth-oauthlib==0.4.1
graphviz==0.13.2
grpcio==1.27.1
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
tensorboard==2.1.0
terminado==0.8.3
testpath==0.4.4
torchviz==0.0.1
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
vvv==0.4.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==2.1.0
